Using a generator function for fetching Data
