+++ src/components/Events/Events.module.scss
@@ -21,7 +21,7 @@
 
 			img {
 				margin-bottom: 3.125rem;
-				margin-left: 100px;
+				margin-right: 100px;
 
 				@media (max-width: 1439px) {
 					width: 16.125rem;
